updates a lot of stuff D

+++ main.py
@@ -2,6 +2,7 @@ import distutils.dir_util
 import glob
 import os
 import shutil
+from pathlib import Path
 
 import cv2
 import deeplabcut
@@ -35,35 +36,52 @@ def get_export_dir(model_name):
     return export_dir
 
 
-def get_dlc_preds(export_dir, src_vid, crop_offset, crop_size, frame_size, frame_is_colored):
-    use_dlc_preprocess = abs(crop_size[0] / frame_size[0] * frame_size[1] - crop_size[1]) < 5
-
-    if use_dlc_preprocess:
-        dlc_live = DLCLive(export_dir,
-                           cropping=[crop_offset[0], crop_offset[0] + crop_size[0],
-                                     crop_offset[1], crop_offset[1] + crop_size[1]],
-                           resize=frame_size[0] / crop_size[0])
-    else:
-        dlc_live = DLCLive(export_dir)
-
-    dlc_live.init_inference(src_vid.read()[1])
-
-    for _ in tqdm(range(int(src_vid.get(cv2.CAP_PROP_FRAME_COUNT)))):
-        ok, frame = src_vid.read()
-        if not ok:
-            break
+def get_dlc_preds(model_name, src_vid, crop_offset, crop_size, frame_size, frame_is_colored, use_live):
+    export_dir = get_export_dir(model_name)
+    if use_live:
+        use_dlc_preprocess = abs(crop_size[0] / frame_size[0] * frame_size[1] - crop_size[1]) < 5
 
         if use_dlc_preprocess:
-            poses = dlc_live.get_pose(frame.astype(np.uint8))
+            dlc_live = DLCLive(export_dir,
+                               cropping=[crop_offset[0], crop_offset[0] + crop_size[0],
+                                         crop_offset[1], crop_offset[1] + crop_size[1]],
+                               resize=frame_size[0] / crop_size[0])
         else:
-            processed_frame = edit_frame(frame, crop_offset, crop_size, frame_size, frame_is_colored)
-            poses = dlc_live.get_pose(processed_frame.astype(np.uint8))
+            dlc_live = DLCLive(export_dir)
+
+        dlc_live.init_inference(src_vid.read()[1])
+
+        for _ in tqdm(range(int(src_vid.get(cv2.CAP_PROP_FRAME_COUNT)))):
+            ok, frame = src_vid.read()
+            if not ok:
+                break
+
+            if use_dlc_preprocess:
+                poses = dlc_live.get_pose(frame.astype(np.uint8))
+            else:
+                processed_frame = edit_frame(frame, crop_offset, crop_size, frame_size, frame_is_colored)
+                poses = dlc_live.get_pose(processed_frame.astype(np.uint8))
+                poses[:, 0] = poses[:, 0] * crop_size[0] / frame_size[0] + crop_offset[0]
+                poses[:, 1] = poses[:, 1] * crop_size[1] / frame_size[1] + crop_offset[1]
+
+            yield frame, poses
+
+        dlc_live.close()
+    else:
+        src_vid, vid_path = src_vid
+        assert frame_size[0] == crop_size[0] and frame_size[1] == crop_size[1]
+        scorer_name = deeplabcut.analyze_videos(get_dlc_config_path(model_name), [vid_path],
+                                                cropping=[crop_offset[0], crop_offset[0] + crop_size[0],
+                                                          crop_offset[1], crop_offset[1] + crop_size[1]])
+        df = pd.read_hdf(f'{Path(vid_path).parents[0]}/{Path(vid_path).stem}{scorer_name}.h5')
+        for i in tqdm(range(int(src_vid.get(cv2.CAP_PROP_FRAME_COUNT)))):
+            ok, frame = src_vid.read()
+            if not ok:
+                break
+            poses = df.loc[i].to_numpy().reshape((-1, 3))
             poses[:, 0] = poses[:, 0] * crop_size[0] / frame_size[0] + crop_offset[0]
             poses[:, 1] = poses[:, 1] * crop_size[1] / frame_size[1] + crop_offset[1]
-
-        yield frame, poses
-
-    dlc_live.close()
+            yield frame, poses
 
 
 def make_program():
@@ -102,26 +120,37 @@ def task_edit_video(configs):
 
 @program_task(2, configurations.main.label_frames)
 def task_label_frames(configs):
+    extract_frames = configs.mode in ['manual', 'automatic']
     for vid in configs.videos:
-        vid_name = f'{vid}_{configs.marker_config.name}'
+        if extract_frames:
+            vid_name = f'{vid}_{configs.marker_config.name}'
+        else:
+            vid_name = f'{Path(configs.mode).stem}_{configs.marker_config.name}'
         video_path = f'{general_configs.videos_dir}/{vid_name}.avi'
         shutil.copy(f'{general_configs.videos_dir}/{vid}.avi', video_path)
+
         create_dlc_project('tmp', configs.experimenter, '.tmp', [video_path], configs.marker_config.markers,
-                           prompt_delete=False, numframes2pick=configs.training_data_frame_count)
+                           prompt_delete=False,
+                           numframes2pick=configs.training_data_frame_count if extract_frames else 1)
         cfg_path = os.path.abspath('.tmp/tmp/config.yaml')
 
         tmp_dir = os.path.abspath(f'.tmp/tmp/labeled-data/{vid_name}')
         dst_dir = os.path.abspath(f'{general_configs.videos_dir}/{general_configs.labels_dir}/{vid_name}')
 
-        deeplabcut.extract_frames(cfg_path, mode=configs.mode, algo='kmeans', crop=False)
+        if extract_frames:
+            deeplabcut.extract_frames(cfg_path, mode=configs.mode, algo='kmeans', crop=False)
+            if os.path.isdir(dst_dir):
+                distutils.dir_util.copy_tree(dst_dir, tmp_dir)
+                shutil.rmtree(dst_dir)
+        else:
+            distutils.dir_util.copy_tree(configs.mode, tmp_dir)
 
-        if os.path.isdir(dst_dir):
-            distutils.dir_util.copy_tree(dst_dir, tmp_dir)
-            shutil.rmtree(dst_dir)
         deeplabcut.label_frames(cfg_path)
 
         shutil.move(tmp_dir, dst_dir)
-        os.remove(video_path)
+
+        if extract_frames:
+            os.remove(video_path)
 
 
 @program_task(3, configurations.main.dlc)
@@ -141,7 +170,7 @@ def task_create_dlc_model(configs):
         for d_name in d_names:
             d_path = f'{general_configs.models_dir}/{configs.model_config.name}/labeled-data/{d_name}'
             shutil.copytree(f'{labels_path}/{d_name}', d_path)
-            for img_path in glob.glob(f'{d_path}/img*.png'):
+            for img_path in glob.glob(f'{d_path}/*.png'):
                 img = np.array(Image.open(img_path))
                 orig_size = img.shape[1::-1]
                 img = edit_frame(img,
@@ -195,8 +224,6 @@ def task_export_model(configs):
 
 @program_task(9, configurations.main.test_system)
 def task_evaluate_accuracy(configs):
-    export_dir = get_export_dir(configs.realtime_config.model_config.name)
-
     cmap = cm.get_cmap('plasma')
     markers = configs.realtime_config.model_config.marker_config.markers
     marker_colors = [np.array(cmap(i / len(markers))[:3]) * 255 for i in range(len(markers))]
@@ -219,8 +246,10 @@ def task_evaluate_accuracy(configs):
                                               int(src_vid.get(cv2.CAP_PROP_FPS)), orig_frame_size, True),
                               f'Cannot create {dst_vid_name}') as dst_vid:
 
-                for frame, poses in get_dlc_preds(export_dir, src_vid, crop_offset, crop_size, frame_size,
-                                                  configs.realtime_config.model_config.frame_config.frame_is_colored):
+                for frame, poses in get_dlc_preds(configs.realtime_config.model_config.name, src_vid,
+                                                  crop_offset, crop_size, frame_size,
+                                                  configs.realtime_config.model_config.frame_config.frame_is_colored,
+                                                  True):
                     for pos, color in zip(poses, marker_colors):
                         prob = pos[2] if pos[2] >= configs.dot_threshold else 0.
                         pos = pos[:2].astype(np.int)
@@ -237,8 +266,6 @@ def task_evaluate_accuracy(configs):
 
 @program_task(10, configurations.main.test_system)
 def task_evaluate_fps(configs):
-    export_dir = get_export_dir(configs.realtime_config.model_config.name)
-
     for vid_name in configs.videos:
         src_vid_name = f'{general_configs.videos_dir}/{vid_name}.avi'
 
@@ -251,9 +278,26 @@ def task_evaluate_fps(configs):
                 configs.realtime_config.model_config.frame_config.frame_size,
                 orig_frame_size)
 
-            for _ in get_dlc_preds(export_dir, src_vid, crop_offset, crop_size, frame_size,
-                                   configs.realtime_config.model_config.frame_config.frame_is_colored):
+            for _ in get_dlc_preds(configs.realtime_config.model_config.name, src_vid, crop_offset, crop_size,
+                                   frame_size, configs.realtime_config.model_config.frame_config.frame_is_colored,
+                                   True):
                 pass
+
+
+@program_task(11, configurations.main.dlc)
+def task_rename_dataset(configs):
+    src = input('name of dataset you want to change: ')
+    dst = input('the new name: ')
+
+    d_path = f'{general_configs.videos_dir}/{general_configs.labels_dir}/{src}'
+
+    dframe = pd.read_hdf(f'{d_path}/CollectedData_{configs.model_config.experimenter}.h5')
+    dframe.index = pd.Index([x.replace(f'labeled-data/{src}', f'labeled-data/{dst}') for x in dframe.index])
+    dframe.to_csv(f'{d_path}/CollectedData_{configs.model_config.experimenter}.csv')
+    dframe.to_hdf(f'{d_path}/CollectedData_{configs.model_config.experimenter}.h5',
+                  'df_with_missing', format='table', mode='w')
+
+    os.rename(d_path, f'{general_configs.videos_dir}/{general_configs.labels_dir}/{dst}')
 
 
 prompt = ''.join([f'{task_num}: {program_task.tasks[task_num].__name__.split("_", 1)[1].replace("_", " ")}\n'
